+++ src/Deserializer/WPP/DecodingContextType.cs
@@ -1,5 +1,8 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System.Collections.ObjectModel;
+using System.Diagnostics.CodeAnalysis;
 using System.Text;
+
+using Nefarius.Utilities.ETW.Deserializer.WPP.TMF;
 
 namespace Nefarius.Utilities.ETW.Deserializer.WPP;
 
@@ -18,9 +21,14 @@ public abstract class DecodingContextType
     protected ReadOnlyMemory<byte> Buffer { get; init; }
 
     /// <summary>
+    ///     Collection of extracted <see cref="TraceMessageFormat"/>s of this <see cref="DecodingContextType"/>.
+    /// </summary>
+    internal abstract ReadOnlyCollection<TraceMessageFormat> TraceMessageFormats { get; }
+
+    /// <summary>
     ///     Managed string representation of <see cref="Buffer" /> content.
     /// </summary>
-    public string BufferValue => Encoding.Unicode.GetString(Buffer.Span[..(Buffer.Length - 2)]);
+    protected string BufferValue => Encoding.Unicode.GetString(Buffer.Span[..(Buffer.Length - 2)]);
 
     /// <summary>
     ///     Turns this instance into a <see cref="TDH_CONTEXT" /> for use with the TDH APIs.
@@ -40,42 +48,6 @@ public abstract class DecodingContextType
                 ParameterSize = 0
             };
         }
-    }
-}
-
-/// <summary>
-///     A <see cref="TDH_CONTEXT_TYPE.TDH_CONTEXT_PDB_PATH" /> wrapper for use with <see cref="DecodingContext" />.
-/// </summary>
-public sealed class PdbFilesDecodingContextType()
-    : DecodingContextType(TDH_CONTEXT_TYPE.TDH_CONTEXT_PDB_PATH)
-{
-    /// <summary>
-    ///     Gets decoding info from one or multiple <c>.PDB</c> files.
-    /// </summary>
-    /// <param name="path">
-    ///     Null-terminated Unicode string that contains the name of the .pdb file for the binary that
-    ///     contains WPP messages. This parameter can be used as an alternative to TDH_CONTEXT_WPP_TMFFILE or
-    ///     TDH_CONTEXT_WPP_TMFSEARCHPATH.
-    /// </param>
-    /// <remarks>To specify multiple files, use <see cref="CreateFrom"/>.</remarks>
-    public PdbFilesDecodingContextType(string path) : this()
-    {
-        ArgumentNullException.ThrowIfNull(path);
-        Buffer = new ReadOnlyMemory<byte>(Encoding.Unicode.GetBytes(Path.GetFullPath(path))
-            .Concat("\0\0"u8.ToArray())
-            .ToArray()
-        );
-    }
-
-    /// <summary>
-    ///     Converts a list of paths to <c>*.pdb</c> files into their corresponding <see cref="PdbFilesDecodingContextType" />
-    ///     objects.
-    /// </summary>
-    /// <param name="pathList">One or more paths to <c>*.pdb</c> files.</param>
-    /// <returns>One or more <see cref="PdbFilesDecodingContextType" />s.</returns>
-    public static IList<DecodingContextType> CreateFrom(params IList<string> pathList)
-    {
-        return pathList.Select(DecodingContextType (path) => new PdbFilesDecodingContextType(path)).ToList();
     }
 }
 
@@ -108,6 +80,9 @@ public sealed class TmfFilesDecodingContextType()
             .ToArray()
         );
     }
+
+    /// <inheritdoc />
+    internal override ReadOnlyCollection<TraceMessageFormat> TraceMessageFormats => throw new NotImplementedException();
 }
 
 /// <summary>
@@ -131,4 +106,7 @@ public sealed class TmfFileDecodingContextType() : DecodingContextType(TDH_CONTE
             .ToArray()
         );
     }
+
+    /// <inheritdoc />
+    internal override ReadOnlyCollection<TraceMessageFormat> TraceMessageFormats => throw new NotImplementedException();
 }
